% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARRAfairness_init.R
\docType{data}
\name{all_data}
\alias{all_data}
\title{All data for fairness measures}
\format{
An object of class \code{list} of length 1260.
}
\usage{
all_data
}
\description{
This object contains all data from analysis of fairness measures in SPARRA v3 and v4.
}
\keyword{data}
\keyword{fairness}
\keyword{sparra}
