% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARRAfairness_functions.R
\name{cal_2panel}
\alias{cal_2panel}
\title{cal_2panel
Draws calibration curves (with legend) with a second panel underneath showing predicted differences.}
\usage{
cal_2panel(
  cals,
  labels,
  col = 1:length(cals),
  xy_col = phs_colours("phs-magenta"),
  ci_col = col,
  highlight = NULL,
  yrange_lower = NULL,
  legend_title = ""
)
}
\arguments{
\item{cals}{list of calibration objects, output from getcal().}

\item{labels}{labels to use in legend}

\item{col}{line colours}

\item{xy_col}{line colour for x-y line, defaults to phs-magenta}

\item{ci_col}{colours to draw confidence intervals on lower panel; NA to not draw.}

\item{highlight}{if non-null, highlight a particular value}

\item{yrange_lower}{y range for lower plot. If NULL, generates automatically}

\item{legend_title}{title for legend, defaults to nothing}
}
\value{
Silently return ggplot object
}
\description{
cal_2panel
Draws calibration curves (with legend) with a second panel underneath showing predicted differences.
}
\examples{
# See vignette
}
