% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARRAfairness_functions.R
\name{logistic}
\alias{logistic}
\title{Logistic}
\usage{
logistic(x)
}
\arguments{
\item{x}{argument}
}
\value{
value of logistic(x)
}
\description{
Logistic function: 1/(1+exp(-x))
}
\examples{

# Plot
x=seq(-5,5,length=1000)
plot(x,logistic(x),type="l")
}
