% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARRAfairness_functions.R
\name{phs_colours}
\alias{phs_colours}
\title{phs_colours}
\usage{
phs_colours(colourname = NULL, keep_names = FALSE)
}
\arguments{
\item{colourname}{name of colour; usually something like phs-blue. If NULL returns all colours.}

\item{keep_names}{keep names of colours in return list. Defaults to false.}
}
\value{
vector of colours, optionally with names.
}
\description{
Copied from github, "Public-Health-Scotland/phsstyles". Public Health Scotland
colour scheme. Internal function.
}
