% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARRAfairness_functions.R
\name{prc_2panel}
\alias{prc_2panel}
\title{prc_2panel
Draws a PRC curve (with legend) with a second panel underneath showing precision difference.}
\usage{
prc_2panel(
  prcs,
  labels = names(prcs),
  col = 1:length(prcs),
  highlight = NULL,
  yrange_lower = NULL,
  legend_title = ""
)
}
\arguments{
\item{prcs}{list of sparraPRC objects.}

\item{labels}{labels to use in legend}

\item{col}{line colours}

\item{highlight}{if non-null, draw a point at a particular cutoff}

\item{yrange_lower}{y range for lower plot. If NULL, generates automatically}

\item{legend_title}{title for legend, defaults to nothing}
}
\value{
Silently return ggplot object
}
\description{
prc_2panel
Draws a PRC curve (with legend) with a second panel underneath showing precision difference.
}
\examples{
# See vignette
}
